AI/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
pymongo
sentence-transformers
faiss-cpu
openai
requests